+++ Changelog.md
@@ -1,4 +1,5 @@
 ### ToyBox Rogue - Ver 2.0.10 (built for 1.6.1.511)
+* (***ADDB***) Add back int parameter to Search 'n Pick
 * (***ADDB***) Units from character picker no longer need to be in-game
 * (***WFMinerva***) Update Chinese Localization (https://github.com/xADDBx/ToyBox-RogueTrader/pull/64)